+++ README.md
@@ -7,12 +7,14 @@ Inspired by the BLEU score matrix shown in Phillip Koehn's [Europarl Statistical
 **NOTE**:
 This codebase is NOT an API nor a Python Package. Type annotation and docstrings were added merely for readability and intepretability's sake, for project evaluators. Only relevant functions/methods and classes received them, whereas private methods or utility function should be understandable by just reading the code (variable names, see how their used, etc.)
 
+## Triplets
+The aligned translations used for evaluation can be found [here](https://drive.google.com/file/d/117-uAof7Us4snuyEZ8yhYfItN0pKn-BH/view?usp=sharing)
+
 ## Translations
 The translations obtained in this project and used for alignment and evaluation can be found [here](https://drive.google.com/file/d/1EJDqTEjVNnDvU59E4sgTSeepkohJin3v/view?usp=sharing)
 * Contains a folder called `translations` which was created with notebook [Preparation.ipynb](https://github.com/na50r/110_bleu/blob/main/Preparation.ipynb)
 
 ## Scores
-The final scores in BLEU, chrF, BERT-F1 and COMET can be found [here](https://drive.google.com/file/d/1TJSvzsO3ZvJB5rbptZrePJ8TRv1wmMcc/view?usp=sharing)
 * We recommend using [RQs.ipynb](https://github.com/na50r/110_bleu/blob/main/RQs.ipynb) as a guideline on how to use the scores to produce tables and visualizations
