--- final_results/ep-deepl.csv
@@ -1,111 +1,0 @@
-Label,BLEU,chrF,BERT-F1,COMET
-da-de,34.06069128304678,62.41010444581509,87.193665,87.75256131589413
-da-el,28.73954655525861,55.2756769722786,85.19641,89.32612949609756
-da-en,40.89269014643669,63.89857118241924,89.16167,86.68762220442295
-da-es,37.1173517383525,61.80483144417068,87.72362,86.29400733858347
-da-fi,25.503814937798108,59.0366073930656,85.48039,91.65195283293724
-da-fr,35.6950815675413,61.59857564829975,87.421326,85.24496058374643
-da-it,27.56729966737891,57.39784495269104,84.79563,86.36169610917568
-da-nl,30.864241133309896,58.32961053695337,85.637474,86.88266751170158
-da-pt,31.047434311962125,57.793132085487,85.890656,85.89748860895634
-da-sv,31.234541702203376,60.26351884066092,86.76156,89.8947735875845
-de-da,34.97438143054963,60.55119110141021,86.935715,89.02560311555862
-de-el,26.949963854458257,54.032105823623056,84.54248,88.92470413446426
-de-en,37.68369750381683,61.78217332643708,88.40994,86.08539512753487
-de-es,36.14866686737648,60.66449242217383,87.00063,85.33908051997423
-de-fi,24.58624963633399,58.34517097625643,84.710335,90.79567262530328
-de-fr,37.80465341371331,63.36534916702978,87.70122,85.67503975331783
-de-it,27.18665113903418,56.63818649108197,84.210526,85.5087053924799
-de-nl,28.33988671191992,56.33514643067743,84.59067,86.75367605686188
-de-pt,30.970712483996827,57.311589004420405,85.70873,85.24406991153955
-de-sv,29.99697480509789,59.15609597137899,86.3492,88.95485427975655
-el-da,34.40131860982505,60.90408473070297,86.790634,89.08693747222424
-el-de,29.992629891097376,61.12701089470303,86.11583,86.41313736140728
-el-en,39.02013829545788,63.7109498298409,88.77196,86.69721788167953
-el-es,38.18625704832871,62.90307495207632,87.624756,86.76800106465816
-el-fi,23.64518259458808,59.17055640792306,85.25218,91.39836084842682
-el-fr,37.2619319728738,63.41971399113603,87.67691,86.34305591881275
-el-it,28.69100470179209,58.52123020325588,85.30872,87.59406633675098
-el-nl,28.52889662335948,57.1249109945438,84.779755,85.69310518354177
-el-pt,34.26269889392885,60.25270773816932,86.53813,86.60471318662167
-el-sv,29.28042186065533,59.11568622223812,85.92711,89.04206496477127
-en-da,37.49732968235935,62.78412168600605,87.497345,90.29563927650452
-en-de,32.54746897306618,61.82734741269461,86.67643,87.92495255172253
-en-el,31.901895715492863,57.5487942950599,85.97888,91.26048858463764
-en-es,41.13749665993079,64.74747891500762,88.39158,88.10633952915668
-en-fi,24.36939823100834,59.30400416440895,85.244705,93.1861815303564
-en-fr,38.49685057200235,64.28522528079814,88.00164,87.79184371232986
-en-it,29.162392605010123,59.32784609169249,85.35848,88.02980102598667
-en-nl,32.20290342196156,59.33410142298564,85.69455,87.69318725168705
-en-pt,34.19596414950691,60.156887359647456,86.21157,87.34387104213238
-en-sv,32.2515683855786,60.797483477378975,86.41182,89.89389905333519
-es-da,37.06557768554858,62.31931291656812,87.38357,88.63704920560122
-es-de,32.86052695892521,62.36606524451148,86.67899,85.65943251922727
-es-el,30.69684758338349,57.03347914674644,85.621735,89.61063335835934
-es-en,43.2742935640392,66.34149700989387,89.496826,86.24395380914211
-es-fi,24.89583671962201,59.63555184246077,85.56644,91.53608018904924
-es-fr,39.91220541501389,64.85493906750133,88.07909,86.35485772788525
-es-it,30.341695464829872,59.97802900074854,85.873634,87.98358158767223
-es-nl,30.271553035710795,57.64925717646027,84.9819,85.31391044706106
-es-pt,35.80915747504449,61.54769165157058,87.146225,87.51502813398838
-es-sv,30.51263395195027,59.80192896090193,86.11908,87.84787089377642
-fi-da,32.6232805913045,58.998924528566135,86.2386,89.84253096580505
-fi-de,28.696167301355167,58.90466750799963,85.487656,86.68714918196201
-fi-el,26.10292382101792,53.19723681075573,84.28041,89.68788078427315
-fi-en,35.48829974263751,60.2097186350415,87.761215,88.01083894819021
-fi-es,32.03020047275304,58.29304523123808,86.393005,86.38052331656218
-fi-fr,32.50616183272121,59.97162540137854,86.58552,86.79962527751923
-fi-it,24.320832205651055,55.37997348057754,83.869995,86.79402287304401
-fi-nl,26.955631319942512,55.23405700669935,84.19033,86.96578979492188
-fi-pt,27.89719650351512,55.731867302439575,85.13358,86.44510126113892
-fi-sv,26.50449356168333,56.88760590578548,85.3007,89.26868129521608
-fr-da,32.87180521194299,59.52418752897694,86.26758,88.25870781391859
-fr-de,30.348471081288828,61.38559360115011,85.93349,86.23386672139168
-fr-el,28.34820543136956,54.94119576655795,84.880615,89.57009017467499
-fr-en,38.30854434425109,62.829230779518106,88.359215,86.09272296726704
-fr-es,37.7039486878334,62.38082324040839,87.4429,87.15665316581726
-fr-fi,22.997571845863888,58.393722410839466,84.99938,91.49323873221874
-fr-it,28.7889797314538,58.78568429278798,85.25447,87.93598414212465
-fr-nl,29.00336680550429,56.71856691448107,84.33731,85.71830306202173
-fr-pt,33.22747626503724,59.504968076714825,86.13452,86.69528175890446
-fr-sv,28.527831711180045,58.31909589023528,85.6517,87.86845372617245
-it-da,29.07291174896692,56.52201025117791,84.6474,87.46686427295208
-it-de,26.528496582318667,56.45887991173064,83.8251,84.86936081200838
-it-el,26.115183810129142,53.336320852395545,83.91058,89.49749165773392
-it-en,33.20450258765978,59.11644977381161,86.65541,84.84649465978146
-it-es,34.74979923267425,60.221307177974495,86.542175,86.86945782601833
-it-fi,19.83369110498071,55.04998484420148,83.14862,90.73000808060168
-it-fr,33.32004581170699,60.47542989319256,86.21264,86.3565051406622
-it-nl,26.3387692014014,55.22137126688741,83.44307,85.08121865987778
-it-pt,29.982963067351253,56.79821128090927,84.69281,86.36490534245968
-it-sv,24.206839514714797,54.8979438353084,83.85693,87.75839675217867
-nl-da,28.717134579940684,55.18875015644152,84.94797,87.94299855828285
-nl-de,25.253817037028245,54.87659826033917,84.123566,85.49018628895283
-nl-el,22.466216291148644,49.2839901074904,82.808975,87.1574135273695
-nl-en,31.239803306485463,56.73403039915499,86.52956,84.4298565685749
-nl-es,29.104524566954176,55.08810661886216,84.98206,83.52793549001217
-nl-fi,19.46410891521297,53.758321973417736,83.20137,90.43516562879086
-nl-fr,29.80866000833105,56.285039212019896,84.99016,83.44881038367748
-nl-it,23.729883013474183,54.19303954204904,83.47864,85.3069448620081
-nl-pt,26.387243417133668,53.69862851402851,83.77301,83.72707960754633
-nl-sv,23.57464156184817,53.98359029016846,83.70797,86.840954862535
-pt-da,32.717717301690314,59.861619888821416,86.340164,88.15422443300486
-pt-de,30.73888874125209,61.04319317155012,86.24948,85.87883690744638
-pt-el,28.862887885472137,55.970466076652514,85.18341,89.45082958042622
-pt-en,37.206175355919314,61.80118538526612,88.051544,85.38264641165733
-pt-es,39.07605900676828,63.54204453954778,87.93707,87.60482946038246
-pt-fi,23.158269389650943,57.879771273243776,84.58673,91.19564262032507
-pt-fr,39.05632413699036,63.85148043337819,87.52971,86.05590775609016
-pt-it,28.814820754131865,58.73866915439538,84.808624,87.33279393613338
-pt-nl,28.687615855491103,56.79186164701865,84.536255,85.24905906617641
-pt-sv,27.93815020317924,57.9737091895914,85.32651,88.06325120478868
-sv-da,36.07028851594553,61.52177465561216,87.38519,90.21843419224024
-sv-de,30.59111734726719,60.647929035484815,86.13986,86.39428237080574
-sv-el,28.03028415477733,54.88792316080786,85.296265,90.00190787017345
-sv-en,39.8606782790255,63.69271988573909,88.91198,86.57975934445858
-sv-es,34.96089524610674,60.14555282673623,86.978195,85.66984852403402
-sv-fi,22.44750726871901,57.06193363280223,84.74004,91.14168958365916
-sv-fr,34.81193817901884,61.17057208927811,87.29435,85.3007095977664
-sv-it,27.391397622045005,57.072334878766775,84.61811,86.86611072719097
-sv-nl,29.28112934266492,56.96685601697211,84.62405,85.85025332868099
-sv-pt,31.222713323062152,58.014357779901005,85.91987,85.96371033787727
