--- final_results/ep-gpt.csv
@@ -1,111 +1,0 @@
-Label,BLEU,chrF,BERT-F1,COMET
-da-de,34.05220657827739,63.10161846745002,87.57031,88.17311254143715
-da-el,27.403131020989917,54.36987097786151,85.33995,89.50874762237072
-da-en,34.26575777834876,60.811370659340746,88.48488,86.3667161166668
-da-es,33.45015942480742,60.48029634424334,87.60773,86.16712042689323
-da-fi,21.559751716183133,56.21152528940844,84.704254,91.14717410504818
-da-fr,31.86284543503208,59.85985607137323,86.02408,85.5570087581873
-da-it,24.794659838043174,55.44263432891824,84.18055,86.18617218100664
-da-nl,26.493696489226185,56.25862439308141,84.8432,86.92690178751945
-da-pt,26.156094816126533,54.65058822378403,85.323944,85.18512783944607
-da-sv,29.32950055719581,59.084340336020034,86.56479,90.02596977353096
-de-da,34.49649539867394,60.963243366346674,87.3864,89.68962541222572
-de-el,23.4013361191418,50.92986565501643,84.1807,88.56253165006638
-de-en,32.61518409570206,59.26927748215392,87.70077,85.83079220982373
-de-es,34.40574046910362,60.05390667262489,86.96981,85.23731433600187
-de-fi,20.38273117116172,54.635433381143685,84.0728,90.66560676860324
-de-fr,31.288818504213022,59.46438121135903,85.45814,85.34031383693218
-de-it,24.302457031725982,54.08144986992277,83.03392,85.26490852981806
-de-nl,25.314172520590837,54.67395952204384,83.73836,86.24405090510845
-de-pt,27.39005329953648,54.75387228757525,85.061874,84.61509845405817
-de-sv,27.25131874158379,57.48022175806218,86.52692,89.48301713168621
-el-da,32.52891361755965,60.26263215817573,86.62122,89.06319697201252
-el-de,27.909468360928592,60.076688692646975,85.60163,86.15073183178902
-el-en,34.10831677377479,61.50316450678508,88.15502,86.66039662063122
-el-es,37.58139415336123,63.59556665863503,88.00008,86.96318717300892
-el-fi,19.81619782649713,55.822753864257656,83.98012,90.21546123921873
-el-fr,36.05842331363316,63.24690674063841,87.64334,86.40258258581161
-el-it,25.871846508542514,57.39000957482377,84.68697,87.53167855920213
-el-nl,25.01885506727986,55.615765245398805,83.97949,85.5498425945451
-el-pt,29.98761899783513,57.96474081182161,86.1243,86.31995785236359
-el-sv,27.174453247463457,58.25509451729727,85.98744,89.49515776336193
-en-da,34.598923583194974,61.08908192824731,86.99877,90.02816705405712
-en-de,27.088721944019778,58.30935577144297,85.62579,87.3743204921484
-en-el,28.5404159965512,55.17998153881936,85.45323,90.81249922513962
-en-es,36.72121720865779,62.623768256043846,88.07785,87.70963659882545
-en-fi,18.983600035991664,55.11494746433381,84.16852,92.7408765034952
-en-fr,33.20795093754573,61.09251635258776,86.21245,87.40726493299007
-en-it,27.096256327915214,57.74327546476439,84.77643,87.60979445278645
-en-nl,27.86802244631609,56.99455928393885,84.81796,87.16294620931149
-en-pt,23.898660906925144,53.31848607978554,84.58525,85.48084910213947
-en-sv,27.72070896361363,58.06046331386761,85.754715,89.65802310407162
-es-da,36.41603774454008,61.96043107552228,87.61045,89.05451959371567
-es-de,32.385555711784846,62.46996357379587,86.76706,86.0038548745215
-es-el,30.406653600175968,57.632084947974015,86.28765,89.98085200032101
-es-en,36.493096981698734,62.820176364929694,88.612656,85.78043279779948
-es-fi,19.21194078359911,54.728078627300434,84.48918,90.96353961527348
-es-fr,35.95266556481712,63.11583045549486,86.702705,86.92706213891506
-es-it,28.261688116762247,58.22996282038486,84.99449,87.2699399292469
-es-nl,27.34848911100012,56.493838388481535,84.36944,85.30808513677573
-es-pt,32.15060096703457,59.3497231506515,86.62705,87.1966468244791
-es-sv,28.04730729891463,57.92792352427523,85.79569,88.0133899897337
-fi-da,28.984256358891383,58.156036711149575,86.01593,89.77741712331772
-fi-de,26.40335048917096,58.11701828366173,85.43935,86.94465736299753
-fi-el,22.93920370291784,51.03439122577888,84.027306,89.67138467729092
-fi-en,31.80530511284869,58.53225021065232,87.491486,87.9073348864913
-fi-es,29.77786422434022,57.89144806370417,86.64397,86.85414003580809
-fi-fr,30.513101100405404,59.23735504893221,86.520584,86.95007292926311
-fi-it,22.02456038043509,54.14725818276811,83.39857,87.56198568392526
-fi-nl,22.120703643113835,53.380605238161706,83.49406,87.16458738604679
-fi-pt,24.91729459476895,54.1401458211242,84.97391,86.47358816117048
-fi-sv,24.41137992092464,56.22611068796084,85.10565,89.59598600119352
-fr-da,33.00770052991709,59.96782855560608,86.60954,88.51894289255142
-fr-de,28.89054234538259,60.07475110887746,85.813736,86.38983577489853
-fr-el,28.0570272276048,55.50196148086596,85.62449,89.926388412714
-fr-en,33.89269339080376,60.61240405754391,87.7444,85.7187360972166
-fr-es,37.49636623211035,63.06123572865469,87.88968,87.20733413100243
-fr-fi,18.709931483403903,54.93158142341578,84.26398,91.2650282084942
-fr-it,27.07845126966881,58.05528169072139,84.693985,87.54511351883411
-fr-nl,27.220359000226225,56.12886541507645,83.986755,85.87075136601925
-fr-pt,27.38379693689822,56.324252510435215,85.59201,85.83833481371403
-fr-sv,27.204276042363738,57.43922654269328,85.67345,88.27507735788822
-it-da,26.915048445732943,56.01773647506749,84.50286,87.89955167472363
-it-de,22.830478310809063,54.86958314623111,82.99839,84.46835895627737
-it-el,24.644441483296227,52.90181417110848,84.1059,89.60587252676487
-it-en,29.14433561595609,57.07379077990195,86.06389,84.44635951519012
-it-es,31.736969311555917,59.39525186714148,86.25255,86.84239296615124
-it-fi,15.083128769750882,51.96729421667317,82.13348,90.21698357183722
-it-fr,28.64636447266793,58.2113969323508,84.26726,86.29127807915211
-it-nl,23.23338605053588,54.45656786557412,82.84301,85.0871029868722
-it-pt,25.929820938156315,55.21220770116292,84.22662,85.98561316728592
-it-sv,21.7341944002217,53.94283972295767,83.746864,87.82176095992327
-nl-da,28.484462739435106,55.38655006661729,85.32271,88.38191932439804
-nl-de,22.56298590856826,53.91266922183205,83.512955,85.48210752755404
-nl-el,21.46406568316484,48.6361526927967,83.17973,87.55510997027159
-nl-en,29.69393780915296,56.23248637876421,86.39124,84.47070814669132
-nl-es,27.445810127766222,54.94436369699185,85.248535,84.13766145706177
-nl-fi,16.06846870861137,51.107225463309526,82.480156,90.32816633582117
-nl-fr,26.946022224740087,54.97433104186194,83.5766,83.86453749984503
-nl-it,22.018027827067133,53.36126276658849,83.43908,85.73664718866348
-nl-pt,22.85577814228849,50.95334937535102,83.39939,83.55297702550888
-nl-sv,22.578141223324288,53.33953136108845,83.76662,87.32474768161774
-pt-da,32.53588404857237,59.42912232661492,86.755486,88.66089801490307
-pt-de,28.399012175458218,60.42673756599988,86.02803,85.96312005072832
-pt-el,28.75130900661551,56.53742872722239,85.742714,89.81765328347683
-pt-en,33.0784985396736,59.84100728920703,87.41196,85.25229692459106
-pt-es,35.746316754130646,63.46122091126233,87.26788,87.4702684879303
-pt-fi,18.49525579870421,53.67432718047059,84.026375,90.77645104838258
-pt-fr,33.79727276071931,62.5758107162229,85.74604,86.2747651040554
-pt-it,25.77190301293807,58.27437215325971,84.26422,87.49477323889732
-pt-nl,25.96131628536167,55.625737094667,83.87892,85.19617177546024
-pt-sv,24.433564583440653,55.77910930673753,85.32329,88.27924098074436
-sv-da,33.09701955266675,61.66924271482821,87.52443,90.63760785758495
-sv-de,27.881856921330836,59.59687413968959,86.351974,87.41410744190216
-sv-el,24.686280627200997,53.02686807752319,85.258286,90.1223549246788
-sv-en,33.80947270581012,60.639145251486056,88.00133,86.0773002281785
-sv-es,32.3720337706934,59.606761963819096,86.89744,85.68677166849375
-sv-fi,19.018249394782757,54.922001000415534,83.72785,91.07968735362068
-sv-fr,30.592760178733915,59.53850655718064,85.71988,85.47852627933025
-sv-it,25.00110116748783,55.96806384397029,84.207,87.01992550492287
-sv-nl,24.14909183387005,54.85654380157487,83.52311,85.73461842536926
-sv-pt,25.22547048629077,54.780163625169976,85.007,85.70261516422033
