--- final_results/flores-deepl.csv
@@ -1,111 +1,0 @@
-Label,BLEU,chrF,BERT-F1,COMET
-da-de,37.90301893526407,64.23223197862603,89.176155,88.41723610460758
-da-el,27.097077397120675,53.56683159270459,86.01889,89.42825335264206
-da-en,54.799530417265686,75.29554272975821,93.8707,91.0443747639656
-da-es,26.36363930725614,53.30465254569262,87.409775,86.06459884345531
-da-fi,26.373986956911292,60.72330283760008,86.66458,92.45266099274158
-da-fr,44.50900771261168,67.30051485152589,90.65958,87.58025686442852
-da-it,30.974264066946343,59.50374413182943,88.28491,88.53248880803585
-da-nl,29.34909929269656,58.84627737024219,86.829346,88.11096984893084
-da-pt,34.96004388966506,61.28746801825887,89.87904,88.19428515015987
-da-sv,39.809583053056485,66.39825135531761,89.79179,91.6548262387514
-de-da,41.11084144245304,65.45929715085688,89.79146,91.1586444452405
-de-el,25.417726775317924,52.22879996542438,85.43617,88.75622445344925
-de-en,49.23013786630629,72.141299190807,92.84284,90.27385644614697
-de-es,24.872797040870985,52.18593212864122,86.810074,85.10124903917313
-de-fi,25.86414525599316,60.114993757547175,86.2193,91.83301584422588
-de-fr,41.51098240758495,65.3561838985263,90.0345,87.14354813098907
-de-it,30.70539775540548,59.11065633477596,88.03444,88.2982846647501
-de-nl,28.574291224242685,58.36222992840445,86.512596,88.50197135657072
-de-pt,33.32566115458266,60.21845532994202,89.3331,87.41190353780985
-de-sv,37.334801021040185,64.49325870496101,89.144226,91.026615858078
-el-da,34.58340374704937,60.690554634024906,88.2125,89.20114488899708
-el-de,31.61539394324072,59.36724417441113,87.378265,86.15772095322609
-el-en,41.84936904281006,67.08944715647452,91.2973,88.86475650966167
-el-es,24.53460490733885,51.2864117637242,86.489456,85.29300674796104
-el-fi,21.63419984167781,57.164911844358166,85.09424,90.7861071676016
-el-fr,38.63548969684878,62.75413567759907,89.11771,86.57710790634155
-el-it,27.993972299261184,56.64076997092994,87.23566,87.57375173270702
-el-nl,24.61994880806996,55.00588698049758,85.42636,86.07535018026829
-el-pt,29.79741079178621,57.6762357626855,88.55861,86.95152051746845
-el-sv,31.002460539245075,59.91659836993536,87.47482,89.22743882238865
-en-da,50.61094748933947,71.82258033958308,92.16862,92.3250334262848
-en-de,44.39383278071737,68.28617191578647,90.42031,89.32145135104656
-en-el,30.64481159302654,56.74062510455322,87.18512,90.60062818229198
-en-es,28.7421676967561,55.50555206497387,88.251915,86.91226477921009
-en-fi,29.7750685318051,63.02593227429602,87.58034,93.57196429371834
-en-fr,52.41387955972797,72.43753043861571,92.39698,89.66738307476044
-en-it,34.821375998576585,62.62445225662808,89.53082,89.77150554955006
-en-nl,32.37245706494853,61.21157350641353,87.628586,89.04396222531796
-en-pt,42.25263758769383,66.51564798837721,91.44636,89.35958358645439
-en-sv,47.07680151419007,71.33405422591429,91.54826,92.24494951963425
-es-da,30.37623078615593,58.40866197993814,87.142815,89.17472660541534
-es-de,27.61803064340313,57.01645926709177,86.611916,85.76426403224468
-es-el,20.553340577336023,48.72638044118592,83.81725,88.29639779031277
-es-en,35.79368370955154,63.63972010627372,90.08158,88.36156029999256
-es-fi,19.90501460932688,55.81608948228601,84.4742,91.24369105696678
-es-fr,35.59361632508074,61.14349292975706,88.53963,86.65337574481964
-es-it,27.18644205569712,56.84995769381052,87.23723,88.51879106462002
-es-nl,24.408719675943747,54.78373487368447,85.2128,86.36085467040539
-es-pt,27.168135610641524,55.763891387953294,87.727066,87.5657132267952
-es-sv,27.571088599508755,57.83456320315361,86.38257,89.00811231136322
-fi-da,33.17410646471886,59.99270058148083,87.69613,90.7914306819439
-fi-de,30.691418168962315,58.903887930565546,87.18446,87.76031358540058
-fi-el,21.96377969319828,49.504108559095485,84.26544,89.2128649353981
-fi-en,38.21398875853247,64.43176436647165,90.60039,90.6009800285101
-fi-es,21.44534343605809,50.018326299655726,85.949646,86.47069858014584
-fi-fr,36.970522959044615,62.25790203251126,88.70932,87.82005093991756
-fi-it,26.393041971707447,55.90385860199305,86.812775,88.77285826206207
-fi-nl,24.79070390118476,55.39185428682502,85.47628,87.96308255940676
-fi-pt,27.87987493367138,56.10619982869069,87.81517,87.9428147226572
-fi-sv,30.1166504944105,59.62469642227162,87.24477,90.86946494877338
-fr-da,38.41685600868511,64.04054431200846,89.39407,90.05118571221828
-fr-de,34.55112974804819,62.2296439424692,88.4155,86.92488914728165
-fr-el,25.43919672318353,52.474449873170045,85.57188,89.24885059893131
-fr-en,49.22721974280475,71.54393280046378,92.94307,90.09216947853564
-fr-es,26.08995581579872,53.180827854997816,87.20914,86.21490056812763
-fr-fi,24.930214603470606,59.19983785405606,86.086426,91.96512986719608
-fr-it,30.602561872910723,59.05979117899053,88.24751,89.11922208964825
-fr-nl,27.320815853950982,56.97279090848853,86.05886,87.15432606637478
-fr-pt,33.634741425314374,60.23175258339886,89.55413,88.00359047949314
-fr-sv,35.32209135127486,63.1499732240919,88.71039,89.80600690841675
-it-da,31.286225008550264,59.61392933859936,87.39588,89.79142202436924
-it-de,29.542371080879192,58.31521421682426,86.892975,86.66395607590675
-it-el,22.166463429460304,49.99516851411644,84.35855,89.06772002577782
-it-en,37.34833413883491,65.03766459149978,90.34651,88.87523491680622
-it-es,24.23786096781189,52.11269546945323,86.50156,86.41696952283382
-it-fi,20.62506636279511,56.63373455858066,84.844925,91.45875695347786
-it-fr,36.91737036547038,62.56941191067109,88.84672,87.77748399972916
-it-nl,23.889912965098265,55.01771551927636,85.08377,86.62669294327497
-it-pt,27.824569215656027,56.7110933291358,87.84056,87.99946366250515
-it-sv,29.05266046151115,59.45682657135192,86.89204,89.76135538518429
-nl-da,31.39327582129465,59.1312592889285,87.135254,89.44091568887234
-nl-de,29.895214216086742,58.27858085924491,86.633026,87.22090277075768
-nl-el,20.48211763402622,48.15198861816538,83.5005,87.33097273856401
-nl-en,36.49095121943854,63.8117484849632,89.77147,88.19920527935028
-nl-es,23.13039768452812,50.08421735711154,85.71996,84.00504188984632
-nl-fi,21.546739304386413,56.79012442708002,84.7999,90.95921829342842
-nl-fr,35.00947702797529,60.81820411570958,88.09041,86.07661792635918
-nl-it,26.15407161098332,55.36246515374808,86.3804,86.88574621081352
-nl-pt,27.54038061197543,55.30300982313312,87.311264,85.84091190993786
-nl-sv,28.197053176612467,58.48910781694848,86.37285,88.97923047095537
-pt-da,39.72365444422841,65.08770007869288,89.92699,90.80633774399756
-pt-de,36.61631080278295,63.14189387080212,88.983154,87.31901393085718
-pt-el,26.289706060206555,53.09130310254352,85.85379,89.40403088927269
-pt-en,53.70148533688752,74.55659996919759,93.645874,90.58234715461732
-pt-es,25.88309240937245,53.15410984177823,87.368164,86.76848861575127
-pt-fi,24.627075458141636,59.0551588258006,86.17987,92.00134743750095
-pt-fr,44.19264179098144,67.11077793430857,90.766365,88.33048294484615
-pt-it,31.00299395399873,59.56826497424327,88.497444,89.35321719944477
-pt-nl,27.102937119581817,57.05462102960688,86.3096,87.25561735033989
-pt-sv,37.07558092538225,64.43340623198436,89.186615,90.60697335004808
-sv-da,43.06670984895605,66.73070214700569,90.31719,91.6703969836235
-sv-de,37.518876704284125,63.29936670529981,88.97615,88.36709189414978
-sv-el,26.767541311963523,53.17361658621677,85.79511,89.44463843107224
-sv-en,53.61012901193646,74.33098465409141,93.53352,90.8391395509243
-sv-es,25.55994558611415,52.50298954915724,86.97664,85.7529454678297
-sv-fi,25.597836710427448,60.60247853684405,86.36163,92.37541519105434
-sv-fr,44.96636996370321,67.09154992870069,90.640915,87.56898175179958
-sv-it,31.13077840312428,59.23213959972535,88.21217,88.65640297532082
-sv-nl,28.536071162275388,57.9071327222876,86.39758,87.70183143764734
-sv-pt,34.9773514017618,60.92900237282609,89.6229,88.10376384854317
