--- final_results/flores-gpt.csv
@@ -1,111 +1,0 @@
-Label,BLEU,chrF,BERT-F1,COMET
-da-de,38.86372343807158,64.65359210992393,89.441864,88.51902677118778
-da-el,25.17720383640351,51.97385567472404,85.664795,89.15912039577961
-da-en,51.67822867027456,74.55116006900734,93.599945,91.02968840301035
-da-es,26.773881784781658,53.814793308873256,87.68477,86.07728759944439
-da-fi,25.40395065303987,59.834235328026494,86.376755,92.283944606781
-da-fr,43.61707468049927,66.97902275176884,90.87678,87.61070597171783
-da-it,29.57951432731664,58.47860386300216,88.06181,88.32170490920544
-da-nl,28.913497392906844,58.83471086761581,86.87117,88.3527310937643
-da-pt,42.01516478751264,65.16971096217551,91.207275,89.6055296510458
-da-sv,38.96863995130944,66.17568215998023,89.55023,91.56700259447098
-de-da,38.69122101933272,63.78995964896711,89.223755,90.70739573985338
-de-el,24.689201897862443,51.042903106298574,85.00113,88.4652249366045
-de-en,48.81194478027723,71.9351961066504,92.95082,90.49801363050938
-de-es,26.09496982183532,52.9071306451054,87.019775,85.29283447563648
-de-fi,25.57226913588357,59.18190836810025,86.15028,91.63380286097528
-de-fr,40.47098565423577,64.78391406866965,89.59055,87.47775429487228
-de-it,29.158139984054,58.2472489455103,87.7376,88.23045720160007
-de-nl,28.88178072640398,58.73785358545206,86.322365,88.44406982511282
-de-pt,38.34922712528269,63.45177519831279,90.175735,88.69643081724644
-de-sv,36.04433060268466,63.50453466273016,88.82483,90.72274842858316
-el-da,34.77340315789733,61.03502200766366,88.33064,89.71972039341927
-el-de,32.386091017439206,60.10074773159272,87.84024,86.66343124210835
-el-en,43.56530609823685,68.19220585863232,91.79853,89.57201500236988
-el-es,25.1883045544156,52.65992097022588,86.95029,85.82275232672691
-el-fi,22.467561938071764,57.20818514161559,85.24193,91.114076718688
-el-fr,39.09559007859208,63.38784596898181,89.4514,86.91400289535522
-el-it,25.62459498147236,56.1555197799747,86.73541,87.9745611846447
-el-nl,26.4456489284964,56.486846363333335,85.44352,86.62264778465033
-el-pt,35.88901312145152,61.10710637750005,89.62443,88.46081832051277
-el-sv,32.38102319292307,60.42672521489722,87.83588,89.68063442409039
-en-da,49.25266328260639,71.4428585199926,92.05684,91.9444833099842
-en-de,43.61430753876413,67.99867454933982,90.86975,89.40231251716614
-en-el,29.047136025747182,55.22801532667712,86.831795,90.3882622718811
-en-es,29.18418203838658,56.26959084221303,88.44646,87.15073144435883
-en-fi,29.36792034215523,62.4407078826575,87.51239,93.40387783944608
-en-fr,51.94997465131191,71.79601801724161,92.52902,89.40991453826427
-en-it,32.68649103010715,61.69027423563883,89.22928,89.6215073466301
-en-nl,30.93016521297787,61.12872616329834,87.67157,88.9339849203825
-en-pt,51.37051195461191,72.09179451262034,93.177986,90.72433932125568
-en-sv,46.35155414198255,70.6160031718856,91.415794,92.05156175792216
-es-da,28.6402570464069,57.52893371613577,86.99113,89.21978744864464
-es-de,28.28704268141224,57.258402053455335,86.74512,86.05271163582802
-es-el,19.83644299093354,47.77035149042442,83.55468,88.38398371636868
-es-en,35.98561313413101,63.9828052013398,90.23325,88.47052109241486
-es-fi,19.249953907553543,55.10622958250278,84.54048,91.15452052652836
-es-fr,32.27095501551716,59.332394499285655,87.44519,86.5284594297409
-es-it,21.66003290304253,53.85323732576685,85.77275,87.79511138796806
-es-nl,23.118785048531883,54.70626438263045,85.13521,86.52972102165222
-es-pt,24.278693425396728,54.95131233330548,87.10348,87.94960828125477
-es-sv,27.26313269420669,57.60615907506518,86.42418,89.00770199298859
-fi-da,30.188756992018615,58.7703731783713,87.27434,90.77162780054849
-fi-de,29.98700198073561,58.98972516118708,87.354355,88.20539574325085
-fi-el,21.104690133841544,48.70195902886847,84.10643,89.46065686643124
-fi-en,39.66020392610423,65.79128347174819,90.79433,91.01956397294998
-fi-es,22.852030753218354,51.088287343369245,86.29781,86.87903636693954
-fi-fr,36.19627395759656,61.95497168548641,88.97991,88.31873098015785
-fi-it,25.794782435380856,55.60957421286633,86.65689,88.924121722579
-fi-nl,25.185672100294692,56.261263883783855,85.85823,88.49359279870987
-fi-pt,32.26625014586176,59.12557684878939,88.94607,89.35227000713348
-fi-sv,27.622318479661505,58.49018865314737,86.89075,90.829394698143
-fr-da,36.655552747390495,62.711961255080936,89.08421,89.62879031896591
-fr-de,34.8756386046337,61.800994471707696,88.681854,87.27604439854622
-fr-el,24.754356031918345,51.40761434992397,85.11604,89.11017727851868
-fr-en,49.69684195137147,71.9325374979775,93.18607,90.2571868300438
-fr-es,26.655845595873743,53.97745972256437,87.36197,86.47984147071838
-fr-fi,24.417396973606277,58.53399976637712,85.974106,91.77354657649994
-fr-it,26.51335023581287,57.15158776217238,87.25604,88.91367557644844
-fr-nl,26.660309194905118,57.44113270615443,86.36173,87.49514501541853
-fr-pt,39.96364410459078,64.39002754702625,90.78397,89.36734208464622
-fr-sv,35.29041917976486,62.58033086733189,88.80594,89.64023306965828
-it-da,29.48884624394942,58.38899786090125,87.08283,89.48944149911404
-it-de,31.180642369488957,59.41517304118958,87.23905,86.61640471220016
-it-el,21.10223576811498,48.97553197262121,84.00199,88.74882981181145
-it-en,39.20655472207818,66.2114377678922,90.738304,89.15247440338135
-it-es,22.636274325759334,51.64344553270919,86.000435,85.93244053423405
-it-fi,20.8523459871046,55.95349051116624,84.60626,91.30747167766094
-it-fr,33.34846518618481,60.85130979565477,88.44133,87.38608630001545
-it-nl,23.42643791616796,55.69146161726378,85.34896,86.85010223835707
-it-pt,29.24424683138493,58.41780829848493,88.10895,88.76918391883373
-it-sv,28.65139868375068,58.7823646553254,86.76344,89.51232728362083
-nl-da,28.519843379988064,57.57748926308045,86.68945,89.17403756082058
-nl-de,28.78649088615137,57.433125220178326,86.35979,86.76676158607006
-nl-el,18.98024781020064,46.76109458440212,83.269165,87.29395717382431
-nl-en,36.99105522640616,64.49189433367054,89.99565,88.61883860826492
-nl-es,23.375163756797072,50.610824518925234,86.0499,84.39978386461735
-nl-fi,18.381937288739827,54.698642853954205,84.180725,90.62124729156494
-nl-fr,33.13107483846575,59.75807929384214,87.378815,86.01704229414463
-nl-it,24.155952038457496,54.441140094087295,86.22056,86.93626229465008
-nl-pt,30.540324821309976,58.25031530147661,88.371346,87.37466350197792
-nl-sv,25.760948362720164,57.474439835303656,86.15317,88.84333664178848
-pt-da,39.13237377809657,64.40081713221859,89.75751,90.6475018709898
-pt-de,36.78797850777234,63.210753101869166,89.03523,87.3958984464407
-pt-el,24.175947844230823,50.96517348853381,85.226944,88.88953655958176
-pt-en,55.54505517939692,75.49218404018818,93.97744,90.87980230152608
-pt-es,25.341997983919256,52.836499056697626,87.070114,86.63675600290298
-pt-fi,24.432898465004083,58.61381226048631,85.9277,91.7338405996561
-pt-fr,44.783242636661626,67.61383176243217,91.28938,88.40646955370903
-pt-it,27.246518890152164,57.211822643960765,87.29931,89.04462276399136
-pt-nl,27.57662194131473,57.835122656105455,86.55747,87.52002234756947
-pt-sv,36.774190249967454,64.04761375547979,89.28698,90.70479902625084
-sv-da,40.109913552315945,65.20075115182561,89.74336,91.38129524886608
-sv-de,38.25017648663524,63.91886259252429,89.26978,88.64769479632378
-sv-el,24.861579230287955,51.54765873078573,85.208595,89.08944924175739
-sv-en,52.92581333409349,74.49479259298879,93.33915,90.7967208623886
-sv-es,26.284242565925275,53.24329153509005,87.08779,85.80430036783218
-sv-fi,24.520846401029843,59.51272482395567,86.00717,91.92275834083556
-sv-fr,41.69300461965706,65.33761183282698,89.74503,87.29384283721447
-sv-it,28.676571966487728,58.17024799451514,87.37369,88.5729022026062
-sv-nl,28.04079174376741,58.18079950584772,86.335754,87.79571820050478
-sv-pt,41.70502419329729,64.90154186902572,90.71289,89.20107516646385
