--- final_results/opus-deepl.csv
@@ -1,21 +1,0 @@
-Label,BLEU,chrF,BERT-F1,COMET
-da-en,40.481974454773045,61.541698976096704,87.47455,83.59948913753033
-de-en,36.40688217969452,57.67900379866128,85.82411,81.45139184594154
-el-en,34.65736586590157,55.42487502488101,86.41727,81.60878500342369
-en-da,38.85667292298236,63.75433033270177,86.64594,85.33443947136402
-en-de,32.448391017464935,55.65210882018511,84.413635,80.90846425294876
-en-el,29.324272774494823,54.09400361878289,84.22906,84.61976756155491
-en-es,39.98642574186635,62.2926428160537,87.25653,82.8599545583129
-en-fi,25.13815153342444,55.83882855317411,81.0423,83.86034985631704
-en-fr,40.30864375121718,63.22562284095971,86.01877,81.06481723487377
-en-it,34.19816906142303,58.27664086724198,84.79237,82.39148830622435
-en-nl,32.199951267176075,55.98310378667082,84.30986,81.1875224635005
-en-pt,34.281772428498925,58.6876603318266,86.51387,82.3886307105422
-en-sv,34.0412442424174,60.783735589334654,84.65263,82.57719324529171
-es-en,44.77857717033263,65.06588901330706,88.80323,84.42909011244774
-fi-en,32.57203380711056,50.90294324142707,83.410385,79.6356487646699
-fr-en,41.57299643017993,63.49138075005726,87.362236,82.90383909642696
-it-en,37.83958743953605,58.3510155518206,86.78775,82.40168108046055
-nl-en,33.96672395122083,54.65322467710448,85.315994,79.9303841367364
-pt-en,40.15885289580548,60.75010028698951,87.92009,83.47794569283724
-sv-en,36.11612096825786,57.276781610968456,85.765625,80.99771887809038
