--- final_results/opus-gpt.csv
@@ -1,21 +1,0 @@
-Label,BLEU,chrF,BERT-F1,COMET
-da-en,37.54064296408391,59.71275717449187,87.397934,83.65445104986429
-de-en,35.06004785782801,57.333422206937314,86.01,81.59852648774782
-el-en,33.19229876631784,54.43151581527489,86.478546,82.30568892508745
-en-da,35.49045281187431,61.3098570502982,86.34947,84.76875618845224
-en-de,30.984189999273998,54.526790956335816,84.495476,80.98096865415573
-en-el,28.44781668263203,52.42070475791312,84.1103,84.11904568225145
-en-es,39.29445637828841,61.691253839539286,87.20652,83.21322737634182
-en-fi,20.24798505187473,51.41751278892121,80.63596,83.37389977276325
-en-fr,35.99483206598813,60.72854551884289,85.1495,81.49482455477118
-en-it,32.2604012760301,57.12556510056227,85.042015,82.71029955148697
-en-nl,30.379753440206564,54.97714399319062,84.04224,81.48383106291294
-en-pt,29.557788068655523,55.33926512785891,85.5082,82.48608278483152
-en-sv,30.547283542676745,58.33456903160437,84.440155,82.32543485611677
-es-en,43.10430025243696,63.976861765137,88.575,84.55484168231487
-fi-en,31.019250674744164,50.67010572120266,83.351906,79.92840276658535
-fr-en,39.515842972669056,62.19833182552048,87.198105,83.02347051352262
-it-en,36.83260424981307,57.88390639471746,87.12454,82.81728972494602
-nl-en,31.038193897983074,53.108779206301904,84.77832,80.87287762761116
-pt-en,37.69407126975156,59.38023399006762,86.91432,83.89464359730482
-sv-en,34.22000979595269,56.37443635568149,85.90674,81.19036207348108
